Remove history package
